refactor: split main logic
